+++ src/libbluray/bdnav/clpi_parse.c
@@ -713,7 +713,7 @@ clpi_parse(const char *path, int verbose)
 
     /* if failed, try backup file */
     if (!cl) {
-        int   len    = strlen(path);
+        size_t len   = strlen(path);
         char *backup = malloc(len + 8);
 
         strncpy(backup, path, len - 18);
